+++ plugins/kubernetes-backend/src/cluster-locator/ConfigClusterLocator.ts
@@ -29,13 +29,32 @@ export class ConfigClusterLocator implements KubernetesClustersSupplier {
     // is required if authProvider is serviceAccount
     return new ConfigClusterLocator(
       config.getConfigArray('clusters').map(c => {
-        return {
+        const authProvider = c.getString('authProvider');
+        const clusterDetails = {
           name: c.getString('name'),
           url: c.getString('url'),
           serviceAccountToken: c.getOptionalString('serviceAccountToken'),
           skipTLSVerify: c.getOptionalBoolean('skipTLSVerify') ?? false,
-          authProvider: c.getString('authProvider'),
+          authProvider: authProvider,
         };
+
+        switch (authProvider) {
+          case 'google': {
+            return clusterDetails;
+          }
+          case 'aws': {
+            const assumeRole = c.getOptionalString('assumeRole');
+            return { assumeRole, ...clusterDetails };
+          }
+          case 'serviceAccount': {
+            return clusterDetails;
+          }
+          default: {
+            throw new Error(
+              `authProvider "${authProvider}" has no config associated with it`,
+            );
+          }
+        }
       }),
     );
   }